Move variables code into chart module
